requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras requirements/requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
deprecated==1.2.14
    # via limits
fastapi==0.114.2
    # via -r requirements/requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via -r requirements/requirements.in
idna==3.9
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.4.5
    # via limits
limits==3.13.0
    # via slowapi
packaging==24.1
    # via limits
pydantic==2.9.1
    # via
    #   -r requirements/requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.3
    # via pydantic
pydantic-settings==2.5.2
    # via -r requirements/requirements.in
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via -r requirements/requirements.in
slowapi==0.1.9
    # via -r requirements/requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.38.5
    # via
    #   -r requirements/requirements.in
    #   fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   limits
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
uvicorn==0.30.6
    # via -r requirements/requirements.in
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
wrapt==1.16.0
    # via deprecated
